20180723-kotlin-bank - AccountService should register a deposit transaction implementation

+++ 20180723-kotlin-bank/src/main/kotlin/Transaction.kt
@@ -2,7 +2,7 @@ import java.time.LocalDate
 
 data class Transaction(
         val date: LocalDate,
-        val amount: Long,
-        val balance: Long
+        val amount: Int,
+        val balance: Int
 )
